+++ src/featureFlags/index.tsx
@@ -2,4 +2,4 @@ export { FeatureFlagsProvider } from "./FeatureFlagsProvider";
 export { EnvVarsStrategy, LocalStorageStrategy } from "./strategies";
 export { useFlag } from "./useFlag";
 export { useAllFlags } from "./useAllFlags";
-export { FlagList } from "./availableFlags";
+export type { FlagList } from "./availableFlags";
